services/fim-api/requirements.txt
watchdog>=4.0.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0
